Rename TextEditor files to ImGuiColorTextEdit
